gathering distributed public structures in a single file
