+++ src/lib/filters/scraps/scrapPing.go
@@ -3,17 +3,8 @@ package scraps
 import (
 	"strconv"
 	"strings"
+	"github.com/zairza-cetb/bench-routes/src/lib/utils"
 )
-
-// TypePingScrap as datatype for ping outputs
-type TypePingScrap struct {
-	Min, Avg, Max, Mdev float64
-}
-
-// TypeFloodPingScrap as datatype for flood ping outputs
-type TypeFloodPingScrap struct {
-	Min, Avg, Max, Mdev, PacketLoss float64
-}
 
 func strToFloat64(s string) float64 {
 	r, err := strconv.ParseFloat(s, 64)
@@ -24,14 +15,14 @@ func strToFloat64(s string) float64 {
 }
 
 // CLIPingScrap scraps the data points for CLIPing function
-func CLIPingScrap(s *string) (a *TypePingScrap) {
+func CLIPingScrap(s *string) (a *utils.TypePingScrap) {
 	arr := strings.Split(*s, "\n")
 	l := len(arr)
 	if l > 2 {
 		words := strings.Split(arr[l-2], " ")
 		if words[0] == "rtt" {
 			temp := strings.Split(words[len(words)-2], "/")
-			a = &TypePingScrap{
+			a = &utils.TypePingScrap{
 				Min:  strToFloat64(temp[0]),
 				Avg:  strToFloat64(temp[1]),
 				Max:  strToFloat64(temp[2]),
@@ -43,7 +34,7 @@ func CLIPingScrap(s *string) (a *TypePingScrap) {
 }
 
 // CLIFLoodPingScrap returns packet loss
-func CLIFLoodPingScrap(s *string) (a *TypeFloodPingScrap) {
+func CLIFLoodPingScrap(s *string) (a *utils.TypeFloodPingScrap) {
 	arr := strings.Split(*s, "\n")
 	sentence := arr[3]
 	words := strings.Split(sentence, ", ")
@@ -51,7 +42,7 @@ func CLIFLoodPingScrap(s *string) (a *TypeFloodPingScrap) {
 
 	pingValuesString := strings.Split(arr[4], " ")[3]
 	pingValues := strings.Split(pingValuesString, "/")
-	a = &TypeFloodPingScrap{
+	a = &utils.TypeFloodPingScrap{
 		Min:        strToFloat64(pingValues[0]),
 		Avg:        strToFloat64(pingValues[1]),
 		Max:        strToFloat64(pingValues[2]),
