+++ src/lib/modules/ping/ping.go
@@ -8,16 +8,16 @@ import (
 	"time"
 )
 
-const (
-	// PathPing stores the defualt address of storage directory of ping data
-	PathPing = "storage/ping"
-)
+// const (
+// 	// PathPing stores the defualt address of storage directory of ping data
+// 	PathPing = "storage/ping"
+// )
 
 // HandlePing is the main handler for ping operations
 func HandlePing(globalChain []*tsdb.ChainPing, urlRaw *string, packets int, tsdbNameHash string, wg *sync.WaitGroup, isTest bool) {
 	chnl := make(chan *string)
 
-	tsdbNameHash = PathPing + "/" + "chunk_ping_" + tsdbNameHash + ".json"
+	tsdbNameHash = utils.PathPing + "/" + "chunk_ping_" + tsdbNameHash + ".json"
 	// launch a goroutine to handle ping operations
 	go utils.CLIPing(urlRaw, packets, chnl)
 	resp := <-chnl
@@ -38,7 +38,7 @@ func HandlePing(globalChain []*tsdb.ChainPing, urlRaw *string, packets int, tsdb
 	wg.Done()
 }
 
-func createNewBlock(val scrap.TypePingScrap) tsdb.BlockPing {
+func createNewBlock(val utils.TypePingScrap) tsdb.BlockPing {
 	return tsdb.BlockPing{
 		Timestamp: time.Now(),
 		Datapoint: tsdb.PingType{
