+++ src/lib/modules/response/resp_delay.go
@@ -17,15 +17,6 @@ const (
 	PathReqResDelay = "storage/req-res-delay-monitoring"
 )
 
-// Response struct
-// This is the object that we return from this module
-// Contains delay in response and the response length
-type Response struct {
-	delay         int
-	resLength     int64
-	resStatusCode int
-}
-
 // HandleResponseDelayForRoute is the initial entrypoint function for this module which takes
 // in a Route struct and supplies it to a function in turn to handle it accordingly. We create
 // channels to run tests for each route in parallel, speeding up the process
@@ -35,7 +26,7 @@ func HandleResponseDelayForRoute(responseChains [][]*tsdb.Chain, route utils.Rou
 	pathDelay := PathReqResDelay + "/" + "chunk_req_res_" + routeSuffix + "_delay.json"
 	pathLength := PathReqResDelay + "/" + "chunk_req_res_" + routeSuffix + "_length.json"
 	pathStatusCode := PathReqResDelay + "/" + "chunk_req_res_" + routeSuffix + "_status.json"
-	c := make(chan Response)
+	c := make(chan utils.Response)
 	go RouteDispatcher(route, c)
 	responseObject := <-c
 	// Store the respective attributes of the
@@ -47,7 +38,7 @@ func HandleResponseDelayForRoute(responseChains [][]*tsdb.Chain, route utils.Rou
 	blockDelay := tsdb.Block{
 		NextBlock:      nil,
 		PrevBlock:      nil,
-		Datapoint:      float32(responseObject.delay),
+		Datapoint:      float32(responseObject.Delay),
 		Timestamp:      time.Now(),
 		NormalizedTime: 0,
 	}
@@ -64,7 +55,7 @@ func HandleResponseDelayForRoute(responseChains [][]*tsdb.Chain, route utils.Rou
 	blockLength := tsdb.Block{
 		NextBlock:      nil,
 		PrevBlock:      nil,
-		Datapoint:      float32(responseObject.resLength),
+		Datapoint:      float32(responseObject.ResLength),
 		Timestamp:      time.Now(),
 		NormalizedTime: 0,
 	}
@@ -81,7 +72,7 @@ func HandleResponseDelayForRoute(responseChains [][]*tsdb.Chain, route utils.Rou
 	blockStatusCode := tsdb.Block{
 		NextBlock:      nil,
 		PrevBlock:      nil,
-		Datapoint:      float32(responseObject.resStatusCode),
+		Datapoint:      float32(responseObject.ResStatusCode),
 		Timestamp:      time.Now(),
 		NormalizedTime: 0,
 	}
@@ -97,20 +88,20 @@ func HandleResponseDelayForRoute(responseChains [][]*tsdb.Chain, route utils.Rou
 }
 
 // RouteDispatcher dispatches a route to respective handlers based on it's request type
-func RouteDispatcher(route utils.Routes, c chan Response) {
+func RouteDispatcher(route utils.Routes, c chan utils.Response) {
 	if route.Method == "GET" {
 		res := HandleGetRequest(route.URL)
 		c <- res
 	} else {
 		// Send a very large integer to automatically rule out as it
 		// is much much larger than the threshold
-		c <- Response{delay: math.MaxInt32, resLength: 0, resStatusCode: 100}
+		c <- utils.Response{Delay: math.MaxInt32,ResLength: 0,ResStatusCode: 100}
 	}
 }
 
 // HandleGetRequest specifically handles routes with GET Requests. Calculates timestamp before
 // and after processing of each request and returns the difference
-func HandleGetRequest(url string) Response {
+func HandleGetRequest(url string) utils.Response {
 	// Time init
 	start := time.Now().UnixNano()
 	resp := utils.SendGETRequest(url)
@@ -121,5 +112,5 @@ func HandleGetRequest(url string) Response {
 	end := time.Now().UnixNano()
 	diff := int((end - start) / int64(time.Millisecond))
 
-	return Response{delay: diff, resLength: resLength, resStatusCode: respStatusCode}
+	return utils.Response{Delay: diff, ResLength: resLength, ResStatusCode: respStatusCode}
 }
